backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "fastapi[all]>=0.116.1",
    "greenlet>=3.2.4",
    "mypy>=1.17.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose>=3.5.0",
    "scalar-fastapi>=1.2.3",
    "sqlmodel>=0.0.24",
    "types-passlib>=1.7.7.20250602",
    "uvicorn[standard]>=0.35.0",
]

[mypy]
ignore_missing_imports = true
